number_of_event() works in a all or nothing way which shouldn't be its behavior. Good progress anyway!

+++ data_generation/structures.hpp
@@ -6,14 +6,18 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 struct genParam {
-	int minSize ;						// Minimum in interval
-	int maxSize ;						// Maximum in interval
-	int typeSection ;					// '0', '1' or '2' depending on the section type
+	size_t minSize ;					// Minimum in interval
+	size_t maxSize ;					// Maximum in interval
+	size_t typeSection ;				// '0', '1' or '2' depending on the section type
 	std::string anchor ;				// Anchor name
 	char typeGen ;						// '+', '*' or none
 	std::vector<std::string> events ;	// List of possible events
 	std::vector<int> attributes ;		// List of events attributes
 } ;
+
+
+////////////////////////////////////////////////////////////////////////////////
+
 
 struct data {
 	std::vector<int> type ;				// '0', '1' or '2' depending on the section type
